LayoutTests/http/tests/download/anchor-download-redirect-same-origin.html

<!DOCTYPE html>
<html>
<head>
<meta name="viewport" content="width=device-width">
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.setShouldLogDownloadCallbacks(true);
        testRunner.waitUntilDownloadFinished();
    }
</script>
</head>
<body>
<p>
Download same origin redirect with anchor in an iframe. Logs should say it works.
<iframe id="cross" src="/download/resources/iframe-anchor-same-origin-redirect-download.html"></iframe>
<p>
</body>
</html>
