LayoutTests/http/tests/download/anchor-download-redirect-same-origin-expected.txt

Download started.
Download was redirected to "http://127.0.0.1:8000/media/resources/test.pdf".
Downloading URL with suggested filename "test.pdf"
Download completed.
Download same origin redirect with anchor in an iframe. Logs should say it works.
